add sql.

+++ api/src/test/resources/schema.sql
@@ -1,1 +1,1 @@
-CREATE TABLE concept_ancestor (ancestor_concept_id bigint, descendant_concept_id bigint, min_levels_of_separation int, max_levels_of_separation int);
+CREATE TABLE concept_ancestor (ancestor_concept_id bigint, descendant_concept_id bigint, min_levels_of_separation bigint, max_levels_of_separation bigint);
